Feature/GT-8 comments

+++ src/main/java/com/gooroomee/api/post/detail/PostDetailController.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
@@ -24,7 +26,7 @@ public class PostDetailController {
     private final BoardRepository boardRepository;
 
     @GetMapping(value = "/{post_id}")
-    public ResponseEntity getPostDetail(@PathVariable(name = "post_id") Long post_id){
+    public ResponseEntity getPostDetail(@PathVariable(name = "post_id") Long post_id) {
         PostDetailDto post = postDetailService.getPost(post_id);
         PostDetailResource postResource = new PostDetailResource(post, post_id);
         postResource.add(linkTo(PostDetailController.class).slash(post_id).withRel("updatePost"));
@@ -36,32 +38,42 @@ public class PostDetailController {
     public ResponseEntity createPost(@RequestBody @Valid PostDetailDto postDetailDto,
                                      @PathVariable(name = "board_type") String boardType,
                                      Errors errors) {
-        try{
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null)
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        try {
             this.postDetailService.storePost(postDetailDto, boardType);
             return ResponseEntity.ok().build();
-        }catch(IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body("게시판등록실패");
         }
     }
 
     @PutMapping("/{post_id}")
     public ResponseEntity updatePost(@PathVariable Long post_id,
-                                      @RequestBody @Valid PostDetailDto postDetailDto) {
-        try{
+                                     @RequestBody @Valid PostDetailDto postDetailDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null)
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        try {
             PostDetailDto updatedPost = this.postDetailService.updatePost(post_id, postDetailDto);
             PostDetailResource postDetailResource = new PostDetailResource(updatedPost, post_id);
             return ResponseEntity.ok(postDetailResource);
-        }catch(IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().build();
         }
     }
 
     @DeleteMapping("/{post_id}")
     public ResponseEntity deletePost(@PathVariable Long post_id) {
-        try{
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null)
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        String visitorId = authentication.getName();
+        try {
             this.postDetailService.deletePost(post_id);
             return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
-        }catch(IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().build();
         }
